requirements_inpaint.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=10.0.0
numpy>=1.24.0
